Refactor import to datastore serializer
