--- src/serializer.ts
@@ -1,5 +1,0 @@
-'use strict';
-
-import datastoreSerializer from './serializers/datastore';
-
-export const Datastore = datastoreSerializer;
