+++ src/serializers/index.ts
@@ -1,0 +1,3 @@
+import datastore from './datastore';
+
+export const datastoreSerializer = datastore;
